:recycle: Move splitVotesIntoRounds to voting utils
